remove or move some functions to fetch.ts
